+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] (1.1.0)
+### Added
+- Non-executable file handling (such files will be open in editor).
+
 ## [1.0.2] - 2022-11-29
 ### Fixed
 - The **Run** tool window no longer get auto-activated on command start.